refactor: move apps/web/config folder to apps/web/src/config
